pyproject.toml
[project]
name = "synthetic-image-generation"
version = "0.1.0"
description = "Synthetic Image Generation tool"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "albumentations==2.0.8",
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "google-genai==1.38.0",
    "numpy==2.2.6",
    "opencv-python-headless==4.12.0.88",
    "pillow==11.3.0",
    "python-dotenv==1.1.1",
    "tqdm==4.67.1",
]
